Waiting Survey List RecyclerView Adapter 수정
