+++ app/src/main/java/com/example/umc_hackathon/post/FormListMarketing.kt
@@ -36,11 +36,11 @@ class FormListMarketing : Fragment(), PostListView {
 
     override fun onGetPostListSuccess(postList: PostListResponse) {
         binding.fragmentMarketingRv.adapter = FormListRAdapter(postList.result)
-        Toast.makeText(activity, "폼 목록을 불러오는데 성공했습니다", Toast.LENGTH_SHORT).show()
+        Log.d("FormListMarketing / ", "마케팅 폼 목록을 불러오는데 성공했습니다")
     }
 
     override fun onGetPostListFailure() {
-        Toast.makeText(activity, "폼 목록을 불러오는데 실패했습니다", Toast.LENGTH_SHORT).show()
+        Log.d("FormListMarketing / ", "마케팅 폼 목록을 불러오는데 실패했습니다")
     }
 
 }
