+++ app/src/main/res/drawable/list_item_style.xml
@@ -8,7 +8,7 @@
             <gradient
                 android:angle="315"
                 android:endColor="@color/white"
-                android:startColor="#1f0f296b" />
+                android:startColor="@color/white" />
             <corners
                 android:radius="20dp" />
             <stroke
